% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_sampledata.R
\docType{data}
\name{FSdata}
\alias{FSdata}
\title{Demo data for latent moderation model}
\format{
A data frame with 403 rows and 9 columns:
\describe{
  \item{ID}{ID}
  \item{DESP1}{Dispositional envy item 1}
  \item{DESP2}{Dispositional envy item 2}
  \item{DESP3}{Dispositional envy item 3}
  \item{UCLA1}{Loneliness item 1}
  \item{UCLA2}{Loneliness item 2}
  \item{AHSP1}{Holistic thinking item 1}
  \item{AHSP2}{Holistic thinking item 2}
  \item{AHSP3}{Holistic thinking item 3}
}
}
\usage{
data(FSdata)
}
\description{
There are 3 latent variables formed by this dataset:
  Dispositional envy (DES) is measured by three indicators: DESP1, DESP2, and DESP3;
  Loneliness (LS) is measured by two indicators: UCLA1 and UCLA2;
  and Holistic thinking (AHS) is measured by three indicators: AHSP1, AHSP2, and AHSP3.
  It is hypothesized that the effect of DES on LS is moderated by AHS.
}
\keyword{datasets}
