% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_sampledata.R
\docType{data}
\name{PISA2012HK}
\alias{PISA2012HK}
\title{PISA 2012 Hong Kong data subset}
\format{
A data frame with 4,670 rows and 21 columns:
\describe{
  \item{SCH_ID}{School ID}
  \item{STD_ID}{Student ID}
  \item{PV1MATH}{Mathematics achievement (plausible value 1)}
  \item{PV1READ}{Reading achievement (plausible value 1)}
  \item{PV1SCIE}{Science achievement (plausible value 1)}
  \item{IMMIG}{Immigrant status. 1 = Native; 2 = 2nd Generation; 3 = 1st Generation}
  \item{FAMSTRUC}{Family structure. 1 = Single parent; 2 = Two parents; 3 = Other}
  \item{FEMALE}{Student gender. 1 = Female; 0 = Male.}
  \item{GRADE}{Student grade}
  \item{ESCS}{Index of economic, social and cultural status}
  \item{HEDRES}{Home educational resources}
  \item{HOMEPOS}{Home possessions}
  \item{CULTPOS}{Cultural possessions}
  \item{MATHEFF}{Mathematics self-efficacy}
  \item{MATINTFC}{Mathematics intentions}
  \item{INSTMOT}{Instrumental motivation for Mathematics}
  \item{INTMAT}{Mathematics interest}
  \item{MATBEH}{Mathematics behaviour}
  \item{MATWKETH}{Mathematics Work Ethic}
  \item{SUBNORM}{Subjective norms in Mathematics}
  \item{PERSEV}{Perseverance}
}
}
\source{
This dataset was retrieved from
  <https://www.oecd.org/pisa/pisaproducts/pisa2012database-downloadabledata.htm>.
  Only 21 variables of the Hong Kong data were extracted and converted to an R dataset.
}
\usage{
data(PISA2012HK)
}
\description{
A subset of data from the PISA 2012 data
}
\keyword{datasets}
