% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_sampledata.R
\docType{data}
\name{SampleData4A}
\alias{SampleData4A}
\title{Sample data for Tutorial Chapter 4 sample output}
\format{
A data frame with 100 rows and 4 columns:
\describe{
  \item{Y}{DV}
  \item{X}{IV}
  \item{M}{Mediator}
  \item{W}{Moderator}
}
}
\usage{
data(SampleData4A)
}
\description{
A dataset for a simple moderated mediation model
}
\keyword{datasets}
