% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_sampledata.R
\docType{data}
\name{SampleData4B}
\alias{SampleData4B}
\title{Sample data for Tutorial Chapter 4 another sample output}
\format{
A data frame with 4,670 rows and 4 columns:
\describe{
  \item{Y}{DV}
  \item{M}{Mediator}
  \item{X}{IV}
  \item{W1}{Moderator 1}
  \item{W2}{Moderator 2}
  \item{W3}{Moderator 3}
  \item{W4}{Moderator 4}
}
}
\usage{
data(SampleData4B)
}
\description{
A dataset for a second-order moderated mediation model
}
\keyword{datasets}
