% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_get_results.R
\name{VS_getCov}
\alias{VS_getCov}
\title{Get covariance matrix}
\usage{
VS_getCov(output = NULL, groupid = "printall")
}
\arguments{
\item{output}{A \code{VS} list object created by \code{\link{VS}} or
\code{\link{VS_groupEqual}}.}

\item{groupid}{A numerical or character variable indicating any group ID for the multi-sample
\code{\link{VS}} output list. If \code{"printall"} (default), a list of covariance matrices
for all groups will be returned. If \code{Multisample} = \code{FALSE}, ignore this argument.}
}
\value{
A covariance matrix of \code{\link{VS}} output data or any group of \code{\link{VS}}
multi-sample output data; or a list of the covariance matrices of all groups of
\code{\link{VS}} multi-sample output data.
}
\description{
Return the covariance matrix of the \code{VS} output data.
}
\examples{
modelspec <- '
 PV1MATH = MATHEFF
 MATHEFF->PV1MATH = ESCS + HEDRES
 HEDRES = ESCS
 '
effectspec <- '
 IV1 = MATHEFF
 DV1 = PV1MATH
 '
# VS model with centered data
VS_model <- VS(PISA2012HK, model = modelspec, effect = effectspec, scale = "center")
# VS multi-sample model with centered data
VS_group <- VS(PISA2012HK, model = modelspec, effect = effectspec, group = "IMMIG",
               scale = "center")
# VS multi-sample model constrainted by group with centered data
VS_group_equal <- VS_groupEqual(VS_group, effect.equal = 2)

# get the covariance matrix for a VS model
cov <- VS_getCov(VS_model)

# get the covariance matrix for a VS multi-sample model with IMMIG = 2
cov <- VS_getCov(VS_group, 2)

# get a list of covariance matrices of all groups for a VS multi-sample model
cov <- VS_getCov(VS_group_equal)

}
