% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_print_results.R
\name{VS_printCor}
\alias{VS_printCor}
\title{Print correlation matrix}
\usage{
VS_printCor(output = NULL, groupid = "printall")
}
\arguments{
\item{output}{A \code{VS} list object created by \code{\link{VS}} or
\code{\link{VS_groupEqual}}.}

\item{groupid}{A numerical or character variable indicating any group ID for the multi-sample
\code{\link{VS}} output list. If \code{"printall"} (default), a list of correlation matrices
for all groups will be returned. If \code{Multisample} = \code{FALSE}, ignore this argument.}
}
\description{
Print the correlation matrix of the \code{VS} output data in \code{VS} style.
}
\examples{
modelspec <- '
 JOYREAD = HEDRES
 HEDRES->JOYREAD = STIMREAD
 STIMREAD = FEMALE
 '
effectspec <- '
 IV1 = HEDRES
 DV1 = JOYREAD
 '
VS_model <- VS(PISA2018HK, model = modelspec, effect = effectspec, scale = "center",
               categorical = "FEMALE")
VS_group <- VS(PISA2018HK, model = modelspec, effect = effectspec, group = "IMMIG",
               scale = "center", categorical = "FEMALE")
VS_group_equal <- VS_groupEqual(VS_group, effect.equal = 2)

# print the correlation matrix for a VS model
VS_printCor(VS_model)

# print the correlation matrix for a VS multi-sample model with IMMIG = 2
VS_printCov(VS_group, 2)

# print the correlation matrices of all groups for a VS multi-sample model
VS_printCor(VS_group_equal)

}
