% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_print_results.R
\name{VS_printMean}
\alias{VS_printMean}
\title{Print mean vector}
\usage{
VS_printMean(output = NULL, groupid = "printall")
}
\arguments{
\item{output}{A \code{VS} list object created by \code{\link{VS}} or
\code{\link{VS_groupEqual}}.}

\item{groupid}{A numerical or character variable indicating any group ID for the multi-sample
\code{\link{VS}} output list. If \code{"printall"} (default), a list of mean vectors
for all groups will be returned. If \code{Multisample} = \code{FALSE}, ignore this argument.}
}
\description{
Print the mean vector of the \code{VS} output data in \code{VS} style.
}
\examples{
modelspec <- '
 PV1MATH = MATHEFF
 MATHEFF->PV1MATH = ESCS + HEDRES
 HEDRES = ESCS
 '
effectspec <- '
 IV1 = MATHEFF
 DV1 = PV1MATH
 '
# VS model with centered data
VS_model <- VS(PISA2012HK, model = modelspec, effect = effectspec, scale = "center")
# VS multi-sample model with centered data
VS_group <- VS(PISA2012HK, model = modelspec, effect = effectspec, group = "IMMIG",
               scale = "center")
# VS multi-sample model constrainted by group with centered data
VS_group_equal <- VS_groupEqual(VS_group, effect.equal = 2)

# print the mean vector for a VS model
VS_printMean(VS_model)

# print the mean vector for a VS multi-sample model with IMMIG = 2
VS_printMean(VS_group, 2)

# print the mean vectors of all groups for a VS multi-sample model
VS_printMean(VS_group_equal)

}
