% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_main.R
\name{VS}
\alias{VS}
\title{Conditional path models by SEM}
\usage{
VS(
  data = NULL,
  model = NULL,
  effect = NULL,
  group = NULL,
  scale = "raw",
  categorical = NULL,
  estimator = "ML",
  bootstrap = 0,
  boot.ci.type = "perc",
  local.dep = TRUE
)
}
\arguments{
\item{data}{A data.frame containing the data to be analyzed.}

\item{model}{A description of the user-specified conceptual model.}

\item{effect}{A description of the user-specified IV and DV for conditional effects.}

\item{group}{A character variable containing the group variable for multi-sample
conditional path models.}

\item{scale}{A character variable containing the scaling method of the data.
Can be \code{"raw"} for raw data (default), \code{"center"} for centered data, or
\code{"std"} for standardized data.}

\item{categorical}{A vector of categorical variable names in data. Categorical
variables will not be centered and standardized even it is requested in the
\code{scale} argument.}

\item{estimator}{A character variable containing the estimator to be used in the
\code{VS} model. Can be \code{"ML"} for maximum likelihood (default), \code{"MLR"}
  = maximum likelihood with ‘HuberWhite’ robust standard errors, or \code{"GLS"} for
  (normal theory) generalized least squares.}

\item{bootstrap}{A numerical value of number of bootstrap draws. If \code{0},
bootstrapping is not used.}

\item{boot.ci.type}{A character variable containing the type of bootstrapping confidence
interval required. Can be \code{"norm"} for first order normal approximation,
\code{"basic"} for basic bootstrap interval, \code{"perc"} for bootstrap percentile
interval (default), or \code{"bca.simple"} for adjusted bootstrap percentile (BCa)
interval with no correction for acceleration (only for bias).}

\item{local.dep}{A logical value indicating whether the covariance between first order
interactions will be added. If value = \code{FALSE}, it is recommended to use
\code{"center"} or \code{"std"} as the scaling method of the data.}
}
\value{
A list containing the information of the \code{VS} model.

\describe{
 \item{\code{Version}}{\code{VS} version.}
 \item{\code{Multisample}}{A logical value indicating whether the output list is a \code{VS}
   multi-sample object.}
 \item{\code{Input}}{A data frame containing the input data.}
 \item{\code{Conceptual}}{A matrix presentation of the \code{VS} conceptual model.}
 \item{\code{Output}}{A data frame containing the output data used in the analysis. Missing
   values are removed by list-wise deletion. Data is centered if \code{scale = "center"} or
   standardized if \code{scale = "std"}.}
 \item{\code{Categorical}}{A vector of character variables containing the names of categorical
   variables}
 \item{\code{Int_labels}}{A data frame storing the information of the interactions created.}
 \item{\code{Scale}}{A character variable indicating the scale of the output data
   (\code{"Raw"}, \code{"Centered"}, \code{"Standardized"}, \code{"Centered by group"}, or
   \code{"Standardized by group"}).}
 \item{\code{Estimator}}{A character variable indicating the estimator to be used in the
   \code{VS} model (\code{"ML"}, \code{"MLR"}, or \code{"GLS"}).}
 \item{\code{Eq_spec}}{A data frame storing the matrix-form equation specification of the
   \code{VS} working model.}
 \item{\code{Cov_spec}}{A data frame storing the matrix-form variance and covariance
   specification of the \code{VS} working model.}
 \item{\code{Order}}{A data frame storing the matrix-form order of moderation of paths
   corresponding to the equation specification.}
 \item{\code{AP}}{A list containing the information of the additional parameters for the
   conditional effects.}
 \item{\code{Group}}{A character variable indicating the group variable for the multi-sample
   conditional path model. If \code{Multisample = FALSE}, value is \code{NULL}.}
 \item{\code{N_groups}}{A numerical value containing the number of groups for the multi-sample
   conditional path model. If \code{Multisample = FALSE}, value is \code{NULL}.}
 \item{\code{GroupID}}{A vector containing the group IDs of the group variable for the
   multi-sample conditional path model. If \code{Multisample = FALSE}, value is \code{NULL}.}
 \item{\code{Constraints}}{A numerical value or a vector of numerical values indicating the group
   equal model constraints set in \code{\link{VS_groupEqual}}. If not created by
   \code{\link{VS_groupEqual}}, value is \code{NULL}.}
 \item{\code{Syntax}}{The \code{lavaan} model syntax of the \code{VS} working model.}
 \item{\code{Model}}{An object of class \code{lavaan} containing the \code{VS} working model.}
 \item{\code{Bootstrap}}{A numerical value indicating the number of bootstrap draws. If \code{0},
   bootstrapping is not used.}
 \item{\code{Seed}}{A numerical value indicating the random seed used for bootstrapping.
   If \code{Bootstrap = 0}, value is \code{NULL}.}
 \item{\code{CItype}}{A character variable containing the type of bootstrapping confidence
   interval required. If \code{Bootstrap = 0}, value is \code{NULL}.}
 \item{\code{CondInfo}}{A data frame storing the IV to DV, Path, Source, and Conditional effect
   columns of the conditional table.}
 \item{\code{CondRemarks}}{A character vector storing the descriptions of the Conditional effects
   and other table remarks.}
 \item{\code{DualStage}}{A list containing the information of dual stage moderated mediation model
   checking.}
}
}
\description{
Convert the user-specified conceptual conditional path model to \code{VS} working
model, run the \code{VS} working model by \code{lavaan}, and store required information
for further result computations.
}
\examples{
modelspec <- '
 JOYREAD = HEDRES
 HEDRES->JOYREAD = STIMREAD
 STIMREAD = FEMALE
 '
effectspec <- '
 IV1 = HEDRES
 DV1 = JOYREAD
 '
VS_model <- VS(PISA2018HK, model = modelspec, effect = effectspec, scale = "center",
               categorical = "FEMALE")
VS_group <- VS(PISA2018HK, model = modelspec, effect = effectspec, group = "IMMIG",
               scale = "center", categorical = "FEMALE")

}
\references{
Joyce Lok Yin Kwan & Wai Chan. (2018). Variable System: An alternative approach for the analysis of
  mediated moderation.  Psychological Methods, 23(2), 262-277.
  \doi{https://doi.org/10.1037/met0000160}
}
