% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_get_results.R
\name{VS_getEffects}
\alias{VS_getEffects}
\title{Get conditional effect table}
\usage{
VS_getEffects(output = NULL, groupid = "printall", boot.ci.type = "perc")
}
\arguments{
\item{output}{A \code{VS} list object created by \code{\link{VS}} or
\code{\link{VS_groupEqual}}.}

\item{groupid}{A numerical or character variable indicating any group ID for the multi-sample
\code{\link{VS}} output list. If \code{"printall"} (default), a list of conditional effect tables
for all groups will be returned. If \code{Multisample} = \code{FALSE}, ignore this argument.}
}
\value{
A conditional effect table of \code{\link{VS}} output data or any group of
\code{\link{VS}} multi-sample output data; or a list of the conditional effect tables
of all groups of \code{\link{VS}} multi-sample output data.
}
\description{
Return the conditional effect table of the \code{VS} output data.
}
\examples{
modelspec <- '
 JOYREAD = HEDRES
 HEDRES->JOYREAD = STIMREAD
 STIMREAD = FEMALE
 '
effectspec <- '
 IV1 = HEDRES
 DV1 = JOYREAD
 '
VS_model <- VS(PISA2018HK, model = modelspec, effect = effectspec, scale = "center",
               categorical = "FEMALE")
VS_group <- VS(PISA2018HK, model = modelspec, effect = effectspec, group = "IMMIG",
               scale = "center", categorical = "FEMALE")
VS_group_equal <- VS_groupEqual(VS_group, effect.equal = 2)

# get the conditional effect table for a VS model
effect <- VS_getEffects(VS_model)

# get the conditional effect table for a VS multi-sample model with IMMIG = 2
effect <- VS_getEffects(VS_group, 2)

# get a list of conditional effect tables of all groups for a VS multi-sample model
effect <- VS_getEffects(VS_group_equal)

}
