% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_groupEqual.R
\name{VS_groupEqual}
\alias{VS_groupEqual}
\title{Multi-sample constrainted conditional path models}
\usage{
VS_groupEqual(output = NULL, effect.equal = NULL)
}
\arguments{
\item{output}{A \code{VS} list object containing the information of the \code{VS} multi-sample
model created by \code{\link{VS}}.}

\item{effect.equal}{A numerical value or a vector of numerical values of the use-specified
group equal constraints chosen from the \code{VS} conceptual model.}
}
\value{
A list containing the information of the \code{VS} model.

\describe{
 \item{\code{Version}}{\code{VS} version.}
 \item{\code{Multisample}}{A logical indicating whether the output list is a \code{VS}
   multi-sample object.}
 \item{\code{Input}}{A data frame containing the input data.}
 \item{\code{Conceptual}}{A matrix presentation of the \code{VS} conceptual model.}
 \item{\code{Output}}{A data frame containing the output data used in the analysis. Missing
   values are removed by list-wise deletion. Data is centered if \code{scale = "center"} or
   standardized if \code{scale = "std"}.}
 \item{\code{Categorical}}{A vector of character variables containing the names of categorical
   variables}
 \item{\code{Int_labels}}{A data frame storing the information of the interactions created.}
 \item{\code{Scale}}{A character variable indicating the scale of the output data
   (\code{"Raw"}, \code{"Centered"}, \code{"Standardized"}, \code{"Centered by group"}, or
   \code{"Standardized by group"}).}
 \item{\code{Estimator}}{A character variable indicating the estimator to be used in the
   \code{VS} model (\code{"ML"}, \code{"MLR"}, or \code{"GLS"}).}
 \item{\code{Eq_spec}}{A data frame storing the matrix-form equation specification of the
   \code{VS} working model.}
 \item{\code{Cov_spec}}{A data frame storing the matrix-form variance and covariance
   specification of the \code{VS} working model.}
 \item{\code{Order}}{A data frame storing the matrix-form order of moderation of paths
   corresponding to the equation specification.}
 \item{\code{AP}}{A list containing the information of the additional parameters for the
   conditional effects.}
 \item{\code{Group}}{A character variable indicating the group variable for the multi-sample
   conditional path model.}
 \item{\code{N_groups}}{A numerical value containing the number of groups for the multi-sample
   conditional path model.}
 \item{\code{GroupID}}{A vector containing the group IDs of the group variable for the
   multi-sample conditional path model.}
 \item{\code{Constraints}}{A numerical value or a vector of numerical values indicating the
   equal group model constraints set.}
 \item{\code{Syntax}}{The \code{lavaan} model syntax of the VS working model.}
 \item{\code{Model}}{An object of class \code{lavaan} containing the \code{VS} multi-sample
   constrainted model.}
 \item{\code{Bootstrap}}{A numerical value indicating the number of bootstrap draws. If \code{0},
   bootstrapping is not used.}
 \item{\code{Seed}}{A numerical value indicating the random seed used for bootstrapping.
   If \code{Bootstrap = 0}, value is \code{NULL}.}
 \item{\code{CItype}}{A character variable containing the type of bootstrapping confidence
   interval required. If \code{Bootstrap = 0}, value is \code{NULL}.}
 \item{\code{CondInfo}}{A data frame storing the IV to DV, Path, Source, and Conditional effect
   columns of the conditional effect table.}
 \item{\code{CondRemarks}}{A character vector storing the descriptions of the Conditional effects
   and other table remarks.}
 \item{\code{DualStage}}{A list containing the information of dual stage moderated mediation model
   checking.}
}
}
\description{
Add user-specified group equal constraints to a \code{VS} multi-sample configural invariance
model, run the \code{VS} multi-sample constrainted model by \code{lavaan}, and store required
information for further result computations.
}
\examples{
modelspec <- '
 JOYREAD = HEDRES
 HEDRES->JOYREAD = STIMREAD
 STIMREAD = FEMALE
 '
effectspec <- '
 IV1 = HEDRES
 DV1 = JOYREAD
 '
VS_group <- VS(PISA2018HK, model = modelspec, effect = effectspec, group = "IMMIG",
               scale = "center", categorical = "FEMALE")
VS_group_equal <- VS_groupEqual(VS_group, effect.equal = 2)

}
