% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_print_results.R
\name{VS_input_summary}
\alias{VS_input_summary}
\title{Print information of input data}
\usage{
VS_input_summary(output = NULL)
}
\arguments{
\item{output}{A \code{VS} list object created by \code{\link{VS}} or
\code{\link{VS_groupEqual}}.}
}
\description{
Print the number of variables, list of variables, and number of observations of the \code{VS}
input data.
}
\examples{
modelspec <- '
 PV1MATH = MATHEFF
 MATHEFF->PV1MATH = ESCS + HEDRES
 HEDRES = ESCS
 '
effectspec <- '
 IV1 = MATHEFF
 DV1 = PV1MATH
 '
VS_model <- VS(PISA2012HK, model = modelspec, effect = effectspec, scale = "center")

VS_input_summary(VS_model)

}
