% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_print_results.R
\name{VS_output_summary}
\alias{VS_output_summary}
\title{Print information of output data}
\usage{
VS_output_summary(output = NULL)
}
\arguments{
\item{output}{A \code{VS} list object created by \code{\link{VS}} or
\code{\link{VS_groupEqual}}.}
}
\description{
Print the number of variables, number of valid cases, data type, list of variables,
and description of interactions created of the \code{VS} output data.
}
\examples{
modelspec <- '
 JOYREAD = HEDRES
 HEDRES->JOYREAD = STIMREAD
 STIMREAD = FEMALE
 '
effectspec <- '
 IV1 = HEDRES
 DV1 = JOYREAD
 '
VS_model <- VS(PISA2018HK, model = modelspec, effect = effectspec, scale = "center",
               categorical = "FEMALE")

VS_output_summary(VS_model)

}
