% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_print_results.R
\name{VS_printParameters}
\alias{VS_printParameters}
\title{Print VS parameter estimates}
\usage{
VS_printParameters(output = NULL, boot.ci.type = NULL)
}
\arguments{
\item{output}{A \code{VS} list object created by \code{\link{VS}} or
\code{\link{VS_groupEqual}}.}

\item{boot.ci.type}{A character variable containing the type of bootstrapping confidence
interval required. Can be \code{"norm"} for first order normal approximation,
\code{"basic"} for basic bootstrap interval, \code{"perc"} for bootstrap percentile
interval (default), or \code{"bca.simple"} for adjusted bootstrap percentile (BCa)
interval with no correction for acceleration (only for bias).}
}
\description{
Print the parameter estimates and confidence intervals of the \code{VS} model.
}
\examples{
modelspec <- '
 PV1MATH = MATHEFF
 MATHEFF->PV1MATH = ESCS + HEDRES
 HEDRES = ESCS
 '
effectspec <- '
 IV1 = MATHEFF
 DV1 = PV1MATH
 '
# VS model with centered data with first order normal approximation bootstrapping C.I.
VS_boot <- VS(PISA2012HK, model = modelspec, effect = effectspec, scale = "center",
              bootstrap = 100, boot.ci.type = "norm")

# print the parameter estimates with model specified C.I.
VS_printParameters(VS_boot)

# print the parameter estimates bootstrap percentile interval
VS_printParameters(VS_boot, "perc")

}
